+++ pyEDAA/IPXACT/Design.py
@@ -36,24 +36,24 @@
 #
 from textwrap           import dedent
 
-from pyIPXACT import RootElement, __DEFAULT_SCHEMA__, Vlnv
+from pyEDAA.IPXACT import RootElement, __DEFAULT_SCHEMA__, Vlnv
 
 
-class DesignConfiguration(RootElement):
-	"""Represents an IP-XACT design configuration."""
+class Design(RootElement):
+	"""Represents an IP-XACT design."""
 
 	def __init__(self, vlnv : Vlnv, description : str):
 		super().__init__(vlnv)
 		
 		self._description =             description
-		self._generatorChainConfiguration =   None
-		self._interconnectionConfiguration =  None
-		self._viewConfiguration =             None
+		self._componentInstances =      []
+		self._interconnections =        []
+		self._adHocConnections =        []
 
-	def SetItem(self, item):
-		if isinstance(item,   GeneratorChainConfiguration):   self._generatorChainConfiguration =   item
-		elif isinstance(item, InterconnectionConfiguration):  self._interconnectionConfiguration =  item
-		elif isinstance(item, ViewConfiguration):             self._viewConfiguration =             item
+	def AddItem(self, item):
+		if isinstance(item, ComponentInstance):   self._componentInstances.append(item)
+		elif isinstance(item, Interconnection):   self._interconnections.append(item)
+		elif isinstance(item, AdHocConnection):   self._adHocConnections.append(item)
 		else:
 			raise ValueError()
 
@@ -62,7 +62,7 @@ class DesignConfiguration(RootElement):
 
 		buffer = dedent("""\
 			<?xml version="1.0" encoding="UTF-8"?>
-			<{xmlns}:designConfiguration
+			<{xmlns}:design
 				xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 				xmlns:{xmlns}="{schemaUri}"
 				xsi:schemaLocation="{schemaUri} {schemaUrl}">
@@ -76,30 +76,54 @@ class DesignConfiguration(RootElement):
 				description=self._description
 			)
 		
-		if self._generatorChainConfiguration:
+		if self._componentInstances:
 			buffer += "\t<{xmlns}:componentInstances>\n"
-			buffer += self._generatorChainConfiguration.ToXml(2)
+			for componentInstance in self._componentInstances:
+				buffer += componentInstance.ToXml(2)
 			buffer += "\t</{xmlns}:componentInstances>\n"
 		
-		if self._interconnectionConfiguration:
-			buffer += "\t<{xmlns}:interconnectionConfiguration>\n"
-			buffer += self._interconnectionConfiguration.ToXml(2)
-			buffer += "\t</{xmlns}:interconnectionConfiguration>\n"
+		if self._interconnections:
+			buffer += "\t<{xmlns}:interconnections>\n"
+			for interconnection in self._interconnections:
+				buffer += interconnection.ToXml(2)
+			buffer += "\t</{xmlns}:interconnections>\n"
 		
-		if self._viewConfiguration:
-			buffer += "\t<{xmlns}:viewConfiguration>\n"
-			buffer += self._viewConfiguration.ToXml(2)
-			buffer += "\t</{xmlns}:viewConfiguration>\n"
+		if self._adHocConnections:
+			buffer += "\t<{xmlns}:adHocConnections>\n"
+			for adHocConnection in self._adHocConnections:
+				buffer += adHocConnection.ToXml(2)
+			buffer += "\t</{xmlns}:adHocConnections>\n"
 		
 		buffer += dedent("""\
-			</{xmlns}:designConfiguration>
+			</{xmlns}:design>
 			""")
 		
 		return buffer.format(xmlns=__DEFAULT_SCHEMA__.NamespacePrefix)
 
 
-class GeneratorChainConfiguration:
-	"""Represents an IP-XACT generator chain configuration."""
+class IpxactFile:
+	def __init__(self, vlnv, name, description):
+		self._vlnv = vlnv
+		self._name = name
+		self._description = description
+	
+	def ToXml(self, indent):
+		"""Converts the object's data into XML format."""
+
+		_indent = "\t" * indent
+		buffer = dedent("""\
+			{indent}<{xmlns}:ipxactFile>
+			{indent}	{vlnv}
+			{indent}	<{xmlns}:name>{path}</{xmlns}:name>
+			{indent}	<{xmlns}:description>{description}</{xmlns}:description>
+			{indent}</{xmlns}:ipxactFile>
+		""").format(indent=_indent, xmlns=__DEFAULT_SCHEMA__.NamespacePrefix, vlnv=self._vlnv.ToXml(0), path=self._name, description=self._description)
+		
+		return buffer
+
+
+class ComponentInstance:
+	"""Represents an IP-XACT component instance."""
 
 	def __init__(self):
 		pass
@@ -110,8 +134,8 @@ class GeneratorChainConfiguration:
 		return ""
 
 
-class InterconnectionConfiguration:
-	"""Represents an IP-XACT interconnection configuration."""
+class Interconnection:
+	"""Represents an IP-XACT interconnection."""
 
 	def __init__(self):
 		pass
@@ -122,8 +146,8 @@ class InterconnectionConfiguration:
 		return ""
 
 
-class ViewConfiguration:
-	"""Represents an IP-XACT view configuration."""
+class AdHocConnection:
+	"""Represents an IP-XACT ad-hoc connection."""
 
 	def __init__(self):
 		pass
