+++ pyEDAA/IPXACT/DesignConfiguration.py
@@ -36,24 +36,24 @@
 #
 from textwrap           import dedent
 
-from pyIPXACT import RootElement, __DEFAULT_SCHEMA__, Vlnv
+from pyEDAA.IPXACT import RootElement, __DEFAULT_SCHEMA__, Vlnv
 
 
-class Design(RootElement):
-	"""Represents an IP-XACT design."""
+class DesignConfiguration(RootElement):
+	"""Represents an IP-XACT design configuration."""
 
 	def __init__(self, vlnv : Vlnv, description : str):
 		super().__init__(vlnv)
 		
 		self._description =             description
-		self._componentInstances =      []
-		self._interconnections =        []
-		self._adHocConnections =        []
+		self._generatorChainConfiguration =   None
+		self._interconnectionConfiguration =  None
+		self._viewConfiguration =             None
 
-	def AddItem(self, item):
-		if isinstance(item, ComponentInstance):   self._componentInstances.append(item)
-		elif isinstance(item, Interconnection):   self._interconnections.append(item)
-		elif isinstance(item, AdHocConnection):   self._adHocConnections.append(item)
+	def SetItem(self, item):
+		if isinstance(item,   GeneratorChainConfiguration):   self._generatorChainConfiguration =   item
+		elif isinstance(item, InterconnectionConfiguration):  self._interconnectionConfiguration =  item
+		elif isinstance(item, ViewConfiguration):             self._viewConfiguration =             item
 		else:
 			raise ValueError()
 
@@ -62,7 +62,7 @@ class Design(RootElement):
 
 		buffer = dedent("""\
 			<?xml version="1.0" encoding="UTF-8"?>
-			<{xmlns}:design
+			<{xmlns}:designConfiguration
 				xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 				xmlns:{xmlns}="{schemaUri}"
 				xsi:schemaLocation="{schemaUri} {schemaUrl}">
@@ -76,54 +76,30 @@ class Design(RootElement):
 				description=self._description
 			)
 		
-		if self._componentInstances:
+		if self._generatorChainConfiguration:
 			buffer += "\t<{xmlns}:componentInstances>\n"
-			for componentInstance in self._componentInstances:
-				buffer += componentInstance.ToXml(2)
+			buffer += self._generatorChainConfiguration.ToXml(2)
 			buffer += "\t</{xmlns}:componentInstances>\n"
 		
-		if self._interconnections:
-			buffer += "\t<{xmlns}:interconnections>\n"
-			for interconnection in self._interconnections:
-				buffer += interconnection.ToXml(2)
-			buffer += "\t</{xmlns}:interconnections>\n"
+		if self._interconnectionConfiguration:
+			buffer += "\t<{xmlns}:interconnectionConfiguration>\n"
+			buffer += self._interconnectionConfiguration.ToXml(2)
+			buffer += "\t</{xmlns}:interconnectionConfiguration>\n"
 		
-		if self._adHocConnections:
-			buffer += "\t<{xmlns}:adHocConnections>\n"
-			for adHocConnection in self._adHocConnections:
-				buffer += adHocConnection.ToXml(2)
-			buffer += "\t</{xmlns}:adHocConnections>\n"
+		if self._viewConfiguration:
+			buffer += "\t<{xmlns}:viewConfiguration>\n"
+			buffer += self._viewConfiguration.ToXml(2)
+			buffer += "\t</{xmlns}:viewConfiguration>\n"
 		
 		buffer += dedent("""\
-			</{xmlns}:design>
+			</{xmlns}:designConfiguration>
 			""")
 		
 		return buffer.format(xmlns=__DEFAULT_SCHEMA__.NamespacePrefix)
 
 
-class IpxactFile:
-	def __init__(self, vlnv, name, description):
-		self._vlnv = vlnv
-		self._name = name
-		self._description = description
-	
-	def ToXml(self, indent):
-		"""Converts the object's data into XML format."""
-
-		_indent = "\t" * indent
-		buffer = dedent("""\
-			{indent}<{xmlns}:ipxactFile>
-			{indent}	{vlnv}
-			{indent}	<{xmlns}:name>{path}</{xmlns}:name>
-			{indent}	<{xmlns}:description>{description}</{xmlns}:description>
-			{indent}</{xmlns}:ipxactFile>
-		""").format(indent=_indent, xmlns=__DEFAULT_SCHEMA__.NamespacePrefix, vlnv=self._vlnv.ToXml(0), path=self._name, description=self._description)
-		
-		return buffer
-
-
-class ComponentInstance:
-	"""Represents an IP-XACT component instance."""
+class GeneratorChainConfiguration:
+	"""Represents an IP-XACT generator chain configuration."""
 
 	def __init__(self):
 		pass
@@ -134,8 +110,8 @@ class ComponentInstance:
 		return ""
 
 
-class Interconnection:
-	"""Represents an IP-XACT interconnection."""
+class InterconnectionConfiguration:
+	"""Represents an IP-XACT interconnection configuration."""
 
 	def __init__(self):
 		pass
@@ -146,8 +122,8 @@ class Interconnection:
 		return ""
 
 
-class AdHocConnection:
-	"""Represents an IP-XACT ad-hoc connection."""
+class ViewConfiguration:
+	"""Represents an IP-XACT view configuration."""
 
 	def __init__(self):
 		pass
